Add authentication to speciality routes

+++ backend/src/controllers/specialtyContoller.ts
@@ -4,20 +4,27 @@ import { Specialty, specialtyDocument } from "../models/specialty";
 export class SpecialtyController {
   constructor() {}
 
-  async createSpecialty(req: Request, res: Response, next: NextFunction) {
+  async createSpecialty(req: any, res: Response, next: NextFunction) {
     try {
+      if (req.role !== "admin") {
+        throw new Error("Unauthorized Operation!");
+      }
       const newSpecialty: specialtyDocument = await new Specialty(req.body);
       await newSpecialty.save();
       res
         .status(201)
         .json({ success: true, msg: "Specialty created succesfully!" });
-    } catch (error) {
+    } catch (error: any) {
+      if (error.msg === "Unautorized Operation!") error.status = 403;
       next(error);
     }
   }
 
-  async updateSpecialty(req: Request, res: Response, next: NextFunction) {
+  async updateSpecialty(req: any, res: Response, next: NextFunction) {
     try {
+      if (req.role !== "admin") {
+        throw new Error("Unauthorized Operation!");
+      }
       const specialty = await Specialty.findByIdAndUpdate(
         req.params.id,
         req.body
@@ -33,8 +40,11 @@ export class SpecialtyController {
     }
   }
 
-  async removeSpecialty(req: Request, res: Response, next: NextFunction) {
+  async removeSpecialty(req: any, res: Response, next: NextFunction) {
     try {
+      if (req.role !== "admin") {
+        throw new Error("Unauthorized Operation!");
+      }
       const specialty = await Specialty.findByIdAndDelete(req.params.id);
       if (!specialty)
         throw new Error("Couldn't find specialty in the database!");
